requirements.txt
flask
requests
Pillow
pytesseract
gunicorn
